requirements.txt
3to2==1.1.1

appdirs==1.4.3
applib==1.2
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
awsebcli==3.14.4
billiard==3.5.0.4
blessed==1.15.0
botocore==1.10.80
cached-property==1.4.3

cement==2.8.2
certifi==2018.1.18

chardet==3.0.4
cmdln==2.0.0
CodeIntel==2.0.0
colorama==0.3.9
constantly==15.1.0

Django==2.0.3

django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
djangorestframework==3.7.7

docopt==0.6.2
docutils==0.14
esprima==4.0.0.dev12
gunicorn==19.8.0

hyperlink==18.0.0
idna==2.6
incremental==17.5.0
Inflector==2.0.12
jmespath==0.9.3
jsonschema==2.6.0
kombu==4.2.1
libclang-py3==3.9.0
mock==2.0.0
msgpack==0.5.6
mysqlclient==1.3.12
numpy==1.14.2
oauthlib==2.1.0

pathspec==0.5.5
pbr==5.2.0
Pillow==5.0.0
python-dateutil==2.7.3
pytz==2018.3
PyYAML==3.13


requests==2.18.4
semantic-version==2.5.0
six==1.11.0
stripe==1.79.1
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.51.0
zope.cachedescriptors==4.3.1
zope.interface==4.4.3
